+++ example/Example.hs
@@ -1,7 +1,8 @@
-{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TemplateHaskell   #-}
 module Main (main) where
 
-import Data.Foldable  (for_)
+import Data.Foldable (for_)
 import FileEmbedLzma (embedByteString, embedRecursiveDir)
 
 import qualified Data.ByteString as BS
@@ -9,5 +10,5 @@ import qualified Data.ByteString as BS
 main :: IO ()
 main = do
     BS.putStr $(embedByteString "example/example.txt")
-    for_ $(embedRecursiveDir "src") $ \(n, bs) ->
+    for_ $(embedRecursiveDir "example") $ \(n, bs) ->
         print (n, BS.length bs)