+++ ChangeLog.md
@@ -3,6 +3,10 @@
 ## 0.0.1
 
 - Use `bytesPrimL` when available
+- Drop GHC-7.x support.
+  We no more depend on `th-lift-instances`,
+  and use `TemplateHaskellQuotes` instead of `TemplateHaskell` for the lib
+  itself.
 
 ## 0